[Extensions] Make CreateJSModuleFromResource public
This function is going to be used by SysApps to register the
common JS API.
